feat: first prototype of the new scrollable mobile UI
